Move card utility functions into card utility module
